+++ CHANGES.rst
@@ -14,7 +14,7 @@ New features:
 
 Bug fixes:
 
-- Fixed unicode-handling of integer values for IChoice. #16
+- Fixed typecasting of IChoice based on its vocabulary. #16
   [iham]
 
 - *add item here*